requirements.txt
# SpeakNode - Python Dependencies

# STT
faster-whisper==1.2.1
torch==2.8.0

# Embedding
sentence-transformers==5.2.2

# LLM orchestration
langchain==1.2.10
langchain-core==1.2.12
langchain-ollama==1.0.1
langchain-community==0.4.1
langgraph==1.0.8

# Graph database
kuzu==0.11.3

# Data modelling
pydantic==2.12.5

# Streamlit UI
streamlit==1.54.0
pyvis==0.3.2
networkx==3.3
matplotlib==3.10.8
Pillow==11.0.0

# FastAPI server
fastapi==0.129.0
uvicorn==0.40.0

# Optional: speaker diarization
# pyannote.audio==3.3.0  # requires HuggingFace token
